Move DownloadFilesView into existing LogsView
